old/ws_receiver/requirements.txt
Pillow>=9.0.0
pillow-heif>=0.7.0
numpy>=1.20.0
astropy>=5.0